+++ pepys_admin/admin_cli.py
@@ -1,105 +1,12 @@
-import argparse
 import cmd
 import os
 from datetime import datetime
 
 from iterfzf import iterfzf
 
-from config import DB_HOST, DB_NAME, DB_PASSWORD, DB_PORT, DB_TYPE, DB_USERNAME
-from pepys_import.core.store.data_store import DataStore
+from pepys_admin.initialise_cli import InitialiseShell
 
 DIR_PATH = os.path.dirname(os.path.abspath(__file__))
-
-
-class InitialiseShell(cmd.Cmd):
-    intro = """--- Menu ---
-(1) Clear database contents
-(2) Clear database schema
-(3) Create Pepys schema
-(4) Import Reference data
-(5) Import Metadata
-(6) Import Sample Measurements
-(0) Exit
-"""
-    prompt = "(initialise) "
-
-    def __init__(self, data_store, parent_shell, csv_path):
-        super(InitialiseShell, self).__init__()
-        self.data_store = data_store
-        self.csv_path = csv_path
-        self.aliases = {
-            "0": self.do_cancel,
-            "1": self.do_clear_db_contents,
-            "2": self.do_clear_db_schema,
-            "3": self.do_create_pepys_schema,
-            "4": self.do_import_reference_data,
-            "5": self.do_import_metadata,
-            "6": self.do_import_sample_measurements,
-        }
-
-        if parent_shell:
-            self.prompt = parent_shell.prompt.strip() + "/" + self.prompt
-
-    def do_clear_db_contents(self):
-        if self.data_store.is_schema_created() is False:
-            return
-
-        self.data_store.clear_db_contents()
-        print("Cleared database contents")
-
-    def do_clear_db_schema(self):
-        if self.data_store.is_schema_created() is False:
-            return
-
-        self.data_store.clear_db_schema()
-        print("Cleared database schema")
-
-    def do_create_pepys_schema(self):
-        self.data_store.initialise()
-        print("Initialised database")
-
-    def do_import_reference_data(self):
-        if self.data_store.is_schema_created() is False:
-            return
-
-        with self.data_store.session_scope():
-            self.data_store.populate_reference(self.csv_path)
-        print("Reference data imported")
-
-    def do_import_metadata(self):
-        if self.data_store.is_schema_created() is False:
-            return
-
-        with self.data_store.session_scope():
-            self.data_store.populate_metadata(self.csv_path)
-        print("Metadata imported")
-
-    def do_import_sample_measurements(self):
-        if self.data_store.is_schema_created() is False:
-            return
-
-        with self.data_store.session_scope():
-            self.data_store.populate_measurement(self.csv_path)
-        print("Sample measurements imported")
-
-    @staticmethod
-    def do_cancel():
-        print("Returning to the previous menu...")
-
-    def default(self, line):
-        cmd_, arg, line = self.parseline(line)
-        if cmd_ in self.aliases:
-            self.aliases[cmd_]()
-            if cmd_ == "0":
-                return True
-        else:
-            print(f"*** Unknown syntax: {line}")
-
-    def postcmd(self, stop, line):
-        if line != "0":
-            print("-" * 61)
-            print(self.intro)
-        return cmd.Cmd.postcmd(self, stop, line)
 
 
 class AdminShell(cmd.Cmd):
@@ -237,24 +144,3 @@ class AdminShell(cmd.Cmd):
             print("-" * 61)
             print(self.intro)
         return cmd.Cmd.postcmd(self, stop, line)
-
-
-def main():
-    parser = argparse.ArgumentParser(description="Pepys Admin CLI")
-    parser.add_argument("--path", type=str, help="CSV files path")
-    args = parser.parse_args()
-
-    data_store = DataStore(
-        db_username=DB_USERNAME,
-        db_password=DB_PASSWORD,
-        db_host=DB_HOST,
-        db_port=DB_PORT,
-        db_name=DB_NAME,
-        db_type=DB_TYPE,
-    )
-
-    AdminShell(data_store, args.path).cmdloop()
-
-
-if __name__ == "__main__":
-    main()
